0.1.3

+++ plugins/functions/program.py
@@ -1,5 +1,5 @@
 # SCP-079-STATUS - Check Linux server status
-# Copyright (C) 2019-2020 SCP-079 <https://scp-079.org>
+# Copyright (C) 2019-2021 SCP-079 <https://scp-079.org>
 #
 # This file is part of SCP-079-STATUS.
 #
@@ -21,20 +21,34 @@ from os import getcwd, getpid, kill
 from signal import SIGABRT
 from subprocess import run
 
+from pyrogram import Client
+
+from .. import glovar
+from .etc import code, lang
+from .telegram import send_message
+
 # Enable logging
 logger = logging.getLogger(__name__)
 
 
-def restart_program() -> bool:
+def restart_program(client: Client, cid: int = 0) -> bool:
     # Restart the program
     result = False
 
+    glovar.locks["restart"].acquire()
+
     try:
+        if cid:
+            send_message(client, cid, f"{lang('status')}：{code(lang('restarted'))}\n")
+            return True
+
         service_name = getcwd().split("/")[-1]
         run(f"systemctl --user restart {service_name}", shell=True)
         kill(getpid(), SIGABRT)
     except Exception as e:
         logger.warning(f"Restart program error: {e}", exc_info=True)
+    finally:
+        glovar.locks["restart"].release()
 
     return result
 
@@ -43,6 +57,8 @@ def update_program() -> bool:
     # Update the program
     result = False
 
+    glovar.locks["restart"].acquire()
+
     try:
         service_name = getcwd().split("/")[-1]
         run(f"bash ~/scp-079/scripts/update.sh {service_name}", shell=True)
@@ -50,5 +66,7 @@ def update_program() -> bool:
         kill(getpid(), SIGABRT)
     except Exception as e:
         logger.warning(f"Update program error: {e}", exc_info=True)
+    finally:
+        glovar.locks["restart"].release()
 
     return result
